lightningd/connect_control: routines to control connecting (move from peer_control.c)
